+++ src/bika/lims/browser/analyses/view.py
@@ -605,6 +605,7 @@ class AnalysesView(ListingView):
         item['Unit'] = format_supsub(obj.getUnit) if obj.getUnit else ''
         item['retested'] = obj.getRetestOfUID and True or False
         item['class']['retested'] = 'center'
+        item['replace']['Service'] = '<strong>{}</strong>'.format(obj.Title)
 
         # Append info link before the service
         # see: bika.lims.site.coffee for the attached event handler
@@ -663,6 +664,8 @@ class AnalysesView(ListingView):
         self._folder_item_fieldicons(obj)
         # Renders remarks toggle button
         self._folder_item_remarks(obj, item)
+        # Renders the analysis conditions
+        self._folder_item_conditions(obj, item)
 
         return item
 
@@ -1377,3 +1380,15 @@ class AnalysesView(ListingView):
             item[position][element] = html
             return
         item[position][element] = glue.join([original, html])
+
+    def _folder_item_conditions(self, analysis_brain, item):
+        """Renders the analysis conditions
+        """
+        analysis = self.get_object(analysis_brain)
+        conditions = analysis.getConditions()
+        if conditions:
+            conditions = map(lambda it: ": ".join([it["title"], it["value"]]),
+                             conditions)
+            conditions = "<br/>".join(conditions)
+            service = item["replace"].get("Service") or item["Service"]
+            item["replace"]["Service"] = "{}<br/>{}".format(service, conditions)